Update nomination_criteria.md
Speaking to the Akatsuki Mania BN team and Akatsuki Nomination Quality Assurance team, it was decided sushi's Mania guidelines were not up to standards. While a new criteria for mania and for the other game modes are discussed to be created, the sushi guidelines have been asked to be removed due to being problematic.